refactor(kubernetes): Move Replacer class to top level
This is a reasonably complex class to be an inner class, which
reduces its ability to encapsulate its complexity. This commit
is just moving the existing class, with no other changes.
