+++ clouddriver-kubernetes/src/main/groovy/com/netflix/spinnaker/clouddriver/kubernetes/v2/artifact/ArtifactReplacer.java
@@ -21,31 +21,22 @@ import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.google.common.collect.ImmutableCollection;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.DocumentContext;
 import com.jayway.jsonpath.JsonPath;
-import com.jayway.jsonpath.PathNotFoundException;
 import com.jayway.jsonpath.spi.json.JacksonJsonNodeJsonProvider;
 import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider;
-import com.netflix.spinnaker.clouddriver.artifacts.kubernetes.KubernetesArtifactType;
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.description.manifest.KubernetesManifest;
 import com.netflix.spinnaker.kork.artifacts.model.Artifact;
 import java.io.IOException;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
-import lombok.Builder;
 import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
@@ -159,104 +150,6 @@ public class ArtifactReplacer {
             })
         .flatMap(Collection::stream)
         .collect(toImmutableSet());
-  }
-
-  @Builder
-  @ParametersAreNonnullByDefault
-  @Slf4j
-  public static class Replacer {
-    private static final ObjectMapper mapper = new ObjectMapper();
-
-    @Nonnull private final String replacePath;
-    @Nonnull private final String findPath;
-    @Nullable private final Function<String, String> nameFromReference;
-    @Nonnull private final KubernetesArtifactType type;
-
-    private static String substituteField(String result, String fieldName, @Nullable String field) {
-      return result.replace("{%" + fieldName + "%}", Optional.ofNullable(field).orElse(""));
-    }
-
-    private static String processPath(String path, Artifact artifact) {
-      String result = substituteField(path, "name", artifact.getName());
-      result = substituteField(result, "type", artifact.getType());
-      result = substituteField(result, "version", artifact.getVersion());
-      result = substituteField(result, "reference", artifact.getReference());
-      return result;
-    }
-
-    private ArrayNode findAll(DocumentContext obj) {
-      return obj.read(findPath);
-    }
-
-    @Nonnull
-    private Artifact artifactFromReference(String s) {
-      return Artifact.builder()
-          .type(type.getType())
-          .reference(s)
-          .name(nameFromReference(s))
-          .build();
-    }
-
-    @Nonnull
-    private String nameFromReference(String s) {
-      if (nameFromReference != null) {
-        return nameFromReference.apply(s);
-      } else {
-        return s;
-      }
-    }
-
-    @Nonnull
-    ImmutableCollection<Artifact> getArtifacts(DocumentContext document) {
-      return mapper
-          .<List<String>>convertValue(findAll(document), new TypeReference<List<String>>() {})
-          .stream()
-          .map(this::artifactFromReference)
-          .collect(toImmutableList());
-    }
-
-    @Nonnull
-    ImmutableCollection<Artifact> replaceArtifacts(
-        DocumentContext obj, Collection<Artifact> artifacts) {
-      ImmutableSet.Builder<Artifact> replacedArtifacts = new ImmutableSet.Builder<>();
-      artifacts.forEach(
-          artifact -> {
-            boolean wasReplaced = replaceIfPossible(obj, artifact);
-            if (wasReplaced) {
-              replacedArtifacts.add(artifact);
-            }
-          });
-      return replacedArtifacts.build();
-    }
-
-    private boolean replaceIfPossible(DocumentContext obj, @Nullable Artifact artifact) {
-      if (artifact == null || StringUtils.isEmpty(artifact.getType())) {
-        throw new IllegalArgumentException("Artifact and artifact type must be set.");
-      }
-
-      if (!artifact.getType().equals(type.getType())) {
-        return false;
-      }
-
-      String jsonPath = processPath(replacePath, artifact);
-
-      log.debug("Processed jsonPath == {}", jsonPath);
-
-      Object get;
-      try {
-        get = obj.read(jsonPath);
-      } catch (PathNotFoundException e) {
-        return false;
-      }
-      if (get == null || (get instanceof ArrayNode && ((ArrayNode) get).size() == 0)) {
-        return false;
-      }
-
-      log.info("Found valid swap for " + artifact + " using " + jsonPath + ": " + get);
-      obj.set(jsonPath, artifact.getReference());
-
-      return true;
-    }
   }
 
   @Value
